fix: improve Docker build stability and test reliability
- Simplify cuda-conda Dockerfile to reduce package installation issues
- Remove unnecessary system packages (git, libglib2.0-0, etc.)
- Use --no-install-recommends to minimize dependencies
- Increase Docker build timeout from 10 to 20 minutes
- Replace jupyter version check with basic container functionality test
- Split test files into test_cuda_conda.py and test_cuda_jupyter.py

These changes address build failures on CI systems and make tests more robust.
